requirements.txt
fastapi
uvicorn[standard]
python-dotenv
SQLAlchemy
pymysql
passlib[bcrypt]   # bcrypt hash/verify
pyjwt
